+++ lib/elixir/test/elixir/regex_test.exs
@@ -40,7 +40,7 @@ defmodule Regex.BinaryTest do
   end
 
   test :unicode do
-    assert ("josé" =~ %r"\p{Latin}$"u) == 3
+    assert ("josé" =~ %r"\p{Latin}$"u)
   end
 
   test :groups do
@@ -93,11 +93,6 @@ defmodule Regex.BinaryTest do
     assert Regex.run(%r"e", "abcd", return: :index) == nil
   end
 
-  test :index do
-    assert Regex.index(%r"c(d)", "abcd") == 2
-    assert Regex.index(%r"e", "abcd") == nil
-  end
-
   test :scan do
     assert Regex.scan(%r"c(d|e)", "abcd abce") == [["d"], ["e"]]
     assert Regex.scan(%r"c(?:d|e)", "abcd abce") == ["cd", "ce"]
@@ -127,6 +122,38 @@ defmodule Regex.BinaryTest do
     assert Regex.replace(%r(b), "abcbe", "[\\&]") == "a[&]c[&]e"
     assert Regex.replace(%r[(b)], "abcbe", "[\\1]") == "a[b]c[b]e"
   end
+
+  test :escape do
+    assert matches_escaped?(".")
+    refute matches_escaped?(".", "x")
+
+    assert matches_escaped?("[\w]")
+    refute matches_escaped?("[\w]", "x")
+
+    assert matches_escaped?("\\")
+
+    assert matches_escaped?("\\xff", "\\xff")
+    refute matches_escaped?("\\xff", "\xff")
+
+    assert matches_escaped?("(")
+    assert matches_escaped?("()")
+    assert matches_escaped?("(?:foo)")
+
+    assert matches_escaped?("\\A  \\z")
+    assert matches_escaped?("  x  ")
+    assert matches_escaped?("  x    x ") # unicode spaces here
+    assert matches_escaped?("# lol")
+
+    assert matches_escaped?("\\A.^$*+?()[{\\| \t\n\xff\\z #hello\x{202F}\x{205F}")
+  end
+
+  defp matches_escaped?(string) do
+    matches_escaped?(string, string)
+  end
+
+  defp matches_escaped?(string, match) do
+    Regex.match? %r/#{Regex.escape(string)}/usimx, match
+  end
 end
 
 defmodule Regex.ListTest do
@@ -149,11 +176,6 @@ defmodule Regex.ListTest do
     assert Regex.run(%r'c(d)', 'abcd') == ['cd', 'd']
     assert Regex.run(%r'e', 'abcd') == nil
     assert Regex.run(%r"c(d)", "abcd", return: :binary) == ["cd", "d"]
-  end
-
-  test :index do
-    assert Regex.index(%r'c(d)', 'abcd') == 2
-    assert Regex.index(%r'e', 'abcd') == nil
   end
 
   test :indexes do