+++ lib/ex_unit/lib/ex_unit/capture_io.ex
@@ -1,7 +1,6 @@
 defmodule ExUnit.CaptureIO do
   @moduledoc %B"""
   This module provides functionality to capture IO to test it.
-  The way to use this module is to import them into your module.
 
   ## Examples
 
@@ -21,12 +20,16 @@ defmodule ExUnit.CaptureIO do
 
   @doc """
   Captures IO. Returns nil in case of no output,
-  otherwise returns the binary which is captured outputs.
+  otherwise returns the binary which is the captured output.
 
   By default, capture_io replaces the group_leader (`:stdio`)
   for the current process. However, the capturing of any other
   named device like `:stderr` is also possible globally by
   giving the registered device name explicitly as argument.
+
+  When capturing `:stdio` and the `:capture_prompt` option is `false`,
+  prompts (specified as arguments in IO.get* functions) are not
+  captured.
 
   A developer can set a string as an input. The default
   input is `:eof`.
@@ -39,36 +42,49 @@ defmodule ExUnit.CaptureIO do
       true
       iex> capture_io(:stderr, fn -> IO.write(:stderr, "josé") end) == "josé"
       true
-      iex> capture_io("this is input", fn->
-      ...>   input = IO.gets ""
+      iex> capture_io("this is input", fn ->
+      ...>   input = IO.gets ">"
+      ...>   IO.write input
+      ...> end) == ">this is input"
+      true
+      iex> capture_io([input: "this is input", capture_prompt: false], fn ->
+      ...>   input = IO.gets ">"
       ...>   IO.write input
       ...> end) == "this is input"
       true
 
   """
-  def capture_io(device, input, fun) do
-    do_capture_io(map_dev(device), input, fun)
+  def capture_io(fun) do
+    do_capture_io(:standard_io, [], fun)
   end
 
   def capture_io(device, fun) when is_atom(device) do
-    do_capture_io(map_dev(device), "", fun)
+    capture_io(device, [], fun)
   end
 
   def capture_io(input, fun) when is_binary(input) do
-    do_capture_io(:standard_io, input, fun)
+    capture_io(:standard_io, [input: input], fun)
   end
 
-  def capture_io(fun) do
-    do_capture_io(:standard_io, "", fun)
+  def capture_io(options, fun) when is_list(options) do
+    capture_io(:standard_io, options, fun)
+  end
+
+  def capture_io(device, input, fun) when is_binary(input) do
+    capture_io(device, [input: input], fun)
+  end
+
+  def capture_io(device, options, fun) when is_list(options) do
+    do_capture_io(map_dev(device), options, fun)
   end
 
   defp map_dev(:stdio),  do: :standard_io
   defp map_dev(:stderr), do: :standard_error
   defp map_dev(other),   do: other
 
-  defp do_capture_io(:standard_io, input, fun) do
+  defp do_capture_io(:standard_io, options, fun) do
     original_gl = :erlang.group_leader
-    capture_gl = new_group_leader(self, input)
+    capture_gl = new_group_leader(self, options)
     :erlang.group_leader(capture_gl, self)
 
     try do
@@ -83,13 +99,15 @@ defmodule ExUnit.CaptureIO do
     end
   end
 
-  defp do_capture_io(device, input, fun) do
+  defp do_capture_io(device, options, fun) do
     unless original_io = Process.whereis(device) do
       raise "could not find IO device registered at #{inspect device}"
     end
 
+    options = Keyword.put(options, :capture_prompt, false)
+
     Process.unregister(device)
-    capture_io = new_group_leader(self, input)
+    capture_io = new_group_leader(self, options)
     Process.register(capture_io, device)
 
     try do
@@ -105,22 +123,26 @@ defmodule ExUnit.CaptureIO do
     end
   end
 
-  defp new_group_leader(runner, input) do
-    spawn_link(fn -> group_leader_process(runner, input) end)
+  defp new_group_leader(runner, options) do
+    spawn_link(fn -> group_leader_process(runner, options) end)
   end
 
-  defp group_leader_process(runner, input) do
+  defp group_leader_process(runner, options) do
+    prompt_config = Keyword.get(options, :capture_prompt, true)
+    input = Keyword.get(options, :input, "")
+
     register_input(input)
+    register_prompt_config(prompt_config)
     group_leader_loop(runner, :infinity, [])
-  end
-
-  defp register_input(nil) do
-    set_input(nil)
   end
 
   defp register_input(input) do
     chars = :unicode.characters_to_list(input)
     set_input(chars)
+  end
+
+  defp register_prompt_config(bool) do
+    Process.put(:capture_io_prompt_config, bool)
   end
 
   defp set_input(:eof) do
@@ -133,6 +155,10 @@ defmodule ExUnit.CaptureIO do
 
   defp get_input do
     Process.get(:capture_io_input)
+  end
+
+  defp need_prompt? do
+    Process.get(:capture_io_prompt_config)
   end
 
   defp group_leader_loop(runner, wait, buf) do
@@ -181,28 +207,42 @@ defmodule ExUnit.CaptureIO do
     io_request({ :put_chars, mod, func, args }, buf)
   end
 
-  defp io_request({ :get_chars, _enc, _prompt, n }, buf) when n >= 0 do
+  defp io_request({ :get_chars, _enc, prompt, n }, buf) when n >= 0 do
+    io_request({ :get_chars, prompt, n }, buf)
+  end
+
+  defp io_request({ :get_chars, prompt, n }, buf) when n >= 0 do
+    if need_prompt? do
+      buf = [prompt|buf]
+    end
+
     { get_chars(n), buf }
   end
 
-  defp io_request({ :get_chars, _prompt, n }, buf) when n >= 0 do
-    { get_chars(n), buf }
+  defp io_request({ :get_line, _enc, prompt }, buf) do
+    io_request({ :get_line, prompt }, buf)
   end
 
-  defp io_request({ :get_line, _prompt }, buf) do
+  defp io_request({ :get_line, prompt }, buf) do
+    if need_prompt? do
+      buf = [prompt|buf]
+    end
+
     { get_line, buf }
   end
 
-  defp io_request({ :get_line, _enc, _prompt }, buf) do
-    { get_line, buf }
+  defp io_request({ :get_until, _encoding, prompt, mod, fun, args}, buf) do
+    io_request({ :get_until, prompt, mod, fun, args}, buf)
   end
 
-  defp io_request({ :get_until, _prompt, mod, fun, args }, buf) do
-    { get_until(mod, fun, args), buf }
-  end
+  defp io_request({ :get_until, prompt, mod, fun, args }, buf) do
+    { result, count } = get_until(mod, fun, args)
 
-  defp io_request({ :get_until, _encoding, _prompt, mod, fun, args}, buf) do
-    { get_until(mod, fun, args), buf }
+    if need_prompt? do
+      buf = [:lists.duplicate(count, prompt)|buf]
+    end
+
+    { result, buf }
   end
 
   defp io_request({ :setopts, _opts }, buf) do
@@ -274,19 +314,19 @@ defmodule ExUnit.CaptureIO do
     do_get_until(input, mod, fun, args)
   end
 
-  defp do_get_until([], mod, fun, args, continuation // [])
+  defp do_get_until([], mod, fun, args, continuation // [], count // 0)
 
-  defp do_get_until([], mod, fun, args, continuation) do
+  defp do_get_until([], mod, fun, args, continuation, count) do
     case apply(mod, fun, [continuation, :eof | args]) do
       { :done, result, rest_chars } ->
         set_input(rest_chars)
-        result
+        { result, count + 1 }
       { :more, next_continuation } ->
-        do_get_until([], mod, fun, args, next_continuation)
+        do_get_until([], mod, fun, args, next_continuation, count + 1)
     end
   end
 
-  defp do_get_until(input, mod, fun, args, continuation) do
+  defp do_get_until(input, mod, fun, args, continuation, count) do
     { line, rest } = Enum.split_while(input, fn(char) -> char != ?\n end)
 
     case rest do
@@ -294,17 +334,17 @@ defmodule ExUnit.CaptureIO do
         case apply(mod, fun, [continuation, line | args]) do
           { :done, result, rest_chars } ->
             set_input(rest_chars)
-            result
+            { result, count + 1 }
           { :more, next_continuation } ->
-            do_get_until([], mod, fun, args, next_continuation)
+            do_get_until([], mod, fun, args, next_continuation, count + 1)
         end
       [_|t] ->
         case apply(mod, fun, [continuation, line ++ '\n' | args]) do
           { :done, result, rest_chars } ->
             set_input(rest_chars ++ t)
-            result
+            { result, count + 1 }
           { :more, next_continuation } ->
-            do_get_until(t, mod, fun, args, next_continuation)
+            do_get_until(t, mod, fun, args, next_continuation, count + 1)
         end
     end
   end