+++ lib/iex/lib/iex/helpers.ex
@@ -99,9 +99,23 @@ defmodule IEx.Helpers do
       h Enum.all?/2
       h Enum.all?
   """
+  # Special case for `h AnyModule.__info__/1`
+  defmacro h({ :/, _, [{ { :., _, [_mod, :__info__] }, _, [] }, 1] }) do
+    quote do
+      IEx.Introspection.h(Module, :__info__, 1)
+    end
+  end
+
   defmacro h({ :/, _, [{ { :., _, [mod, fun] }, _, [] }, arity] }) do
     quote do
       IEx.Introspection.h(unquote(mod), unquote(fun), unquote(arity))
+    end
+  end
+
+  # Special case for `h AnyModule.__info__`
+  defmacro h({ { :., _, [_mod, :__info__] }, _, [] }) do
+    quote do
+      IEx.Introspection.h(Module, :__info__, 1)
     end
   end
 
@@ -233,7 +247,7 @@ defmodule IEx.Helpers do
   in the current IEx session.
   """
   def r do
-    Enum.map iex_reloaded, r(&1)
+    List.flatten(Enum.map(iex_reloaded, do_r(&1)))
   end
 
   @doc """
@@ -243,16 +257,24 @@ defmodule IEx.Helpers do
   are recompiled and reloaded.
   """
   def r(module) do
+    case do_r(module) do
+      mods when is_list(mods) -> { module, mods }
+      other -> other
+    end
+  end
+
+  defp do_r(module) do
     if source = source(module) do
       Process.put(:iex_reloaded, :ordsets.add_element(module, iex_reloaded))
-      { module, Code.load_file source }
+      Enum.map(Code.load_file(source), fn {name, _} -> name end)
     else
       :nosource
     end
   end
 
   @doc """
-  Purges and reloads specified module.
+  Loads given module beam code (and ensures any previous
+  old version was properly purged before).
   """
   def l(module) do
     :code.purge(module)
@@ -285,8 +307,8 @@ defmodule IEx.Helpers do
     source = module.module_info(:compile)[:source]
 
     case source do
-      { :source, source } -> list_to_binary(source)
-      _ -> nil
+      nil -> nil
+      source -> list_to_binary(source)
     end
   end