+++ lib/mix/test/mix/tasks/deps_test.exs
@@ -106,7 +106,7 @@ defmodule Mix.Tasks.DepsTest do
     Mix.Project.push OutOfDateDepsApp
 
     in_fixture "deps_status", fn ->
-      assert_raise Mix.OutOfDateDepsError, "Some dependencies are out of date, please run `MIX_ENV=dev mix deps.get` to proceed", fn ->
+      assert_raise Mix.OutOfDateDepsError, fn ->
         Mix.Tasks.Deps.Check.run []
       end
 
@@ -287,6 +287,9 @@ defmodule Mix.Tasks.DepsTest do
       message = "* Getting git_repo [git: #{inspect fixture_path("git_repo")}]"
       assert_received { :mix_shell, :info, [^message] }
       assert_received { :mix_shell, :info, ["Generated git_repo.app"] }
+
+      # Make sure retriever uses converger
+      refute_received { :mix_shell, :info, [^message] }
 
       Mix.Task.clear
       Mix.Tasks.Deps.Update.run []